third_party/WebKit/LayoutTests/accessibility/selection-states-expected.txt

Option 1
Option 2
Option 3

This tests that we report the correct selection-related states.

PASS axSelectElement.isMultiSelectable is true
PASS axSelectElement.childAtIndex(0).isSelectable is true
PASS axSelectElement.childAtIndex(0).isSelected is true
PASS axSelectElement.childAtIndex(1).isSelectable is true
PASS axSelectElement.childAtIndex(1).isSelected is false
PASS axSelectElement.childAtIndex(2).isSelectable is false
PASS axSelectElement.childAtIndex(2).isSelected is false
PASS successfullyParsed is true

TEST COMPLETE
